Governance
The community is free to choose its own targets and objectives. That said, our goal is to Develop apps on blockchain.

NO 1
There will be a council always composed of 3 members. They are the Council team in our organization.

NO 2
There will be a high council composed of 1 member of the council. Will be the Maintainers of the Council team.

NO 3
Any member of GD can create and collaborate on projects.

NO 4
There will be  weekly reports .

NO 5
In case of any support from reward or bounty, the funds will initially be held :-

For Eos Virtual Hackathon Competition its EOS wallet (it will be public and transaction can be traced) for those  
Wallet Address:{Address will be provided soon}


NO 6
all GD funds will be held in a multi-sig wallet requiring a majority of the high council to unlock funding.

NO 7
Rewards distribution will be voted on by project before being sent. There must be a majority agreement of all 3 council members.

NO 8
All decisions not explicitly differentiated in this list will be voted upon as in item 7.

NO 9
Item 7 voting polls of the council will run for 1 week on the #council Discord channel and after closing will be published in the #develop channel.

NO 10
The high council will be voted upon each quarter, synchronizing with the quarters of the year. The first election will be on 1 October 2017. The newly elected high council can propose changes to these rules during the first week after the election. Changes will be voted upon as in item 7.

NO 11
Any member of the council or high council is free to forfeit their position at any time. When this happens, an election will take place and a majority of council members must decide upon a replacement.

NO 12
Any member of the council that does not vote on polls for more than 2 months will automatically forfeit their position.